Add ability to save and load bubble schemes
Add menu with 3 fields - save, load, exit. Now save function search at
home directory path document/document and try save here file *.xml. Than
you can load this save file.

+++ bubble.cpp
@@ -2,7 +2,7 @@
 
 using namespace std;
 
-Bubble::Bubble(int* id, QPoint* position, QColor* color, QString* name, float* size) {
+Bubble::Bubble(int* id, QPoint* position, QColor* color, QString* name, double* size) {
     SetPosition(position);
     SetColor(color);
     SetDefaultColor(new QColor(112, 122, 116));
@@ -62,11 +62,11 @@ void Bubble::SetName(QString* name) {
     _name = name;
 }
 
-float* Bubble::GetBubbleSize() {
+double* Bubble::GetBubbleSize() {
     return _size;
 }
 
-void Bubble::SetBubbleSize(float* size) {
+void Bubble::SetBubbleSize(double* size) {
     _size = size;
 }
